requirements.txt
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
requests==2.32.3
types-psycopg2==2.9.21.20240819
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
